pyproject.toml
[project]
name = "raspberry-pi-color-sensor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "pyzmq>=27.1.0",
    "rpi-gpio==0.7.1",
]
